refactor(agents): consolidate prompt-phase attempt steps
