+++ qa/scenarios/runtime/cjk-tool-result-budget.yaml
@@ -18,7 +18,7 @@ scenario:
   codeRefs:
     - src/agents/embedded-agent-runner/tool-result-text-budget.ts
     - src/agents/embedded-agent-runner/tool-result-truncation.ts
-    - src/agents/embedded-agent-runner/run/attempt-prompt-context.ts
+    - src/agents/embedded-agent-runner/run/attempt-prompt-build.ts
   execution:
     kind: flow
     summary: Read dense CJK through the real agent loop and inspect the provider-bound truncated tool result.
