+++ src/agents/embedded-agent-runner/run/attempt-prompt-execution-prepare.test.ts
@@ -10,14 +10,15 @@ vi.mock("@openclaw/media-core/constants", () => ({
   mediaKindFromMime: (mime?: string) =>
     mime ? (mime.startsWith("image/") ? "image" : "unknown") : undefined,
 }));
-vi.mock("../../image-sanitization.js", () => ({
+vi.mock("../../image-sanitization.js", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("../../image-sanitization.js")>()),
   resolveImageSanitizationLimits: hoisted.resolveImageSanitizationLimits,
 }));
 vi.mock("./images.js", () => ({
   detectAndLoadPromptImages: hoisted.detectAndLoadPromptImages,
 }));
 
-import { prepareEmbeddedAttemptPromptExecution } from "./attempt-prompt-execution-prepare.js";
+import { prepareEmbeddedAttemptPromptExecution } from "./attempt-prompt-submit.js";
 
 type PromptExecutionInput = Parameters<typeof prepareEmbeddedAttemptPromptExecution>[0];
