--- src/agents/embedded-agent-runner/run/attempt-prompt-execution-prepare.ts
@@ -1,71 +1,0 @@
-/** Prepares prompt-lock ownership and prompt-local images for submission. */
-import { MAX_IMAGE_BYTES } from "@openclaw/media-core/constants";
-import { readPersistedMediaFacts } from "../../../media/media-facts.js";
-import { resolveImageSanitizationLimits } from "../../image-sanitization.js";
-import type { SandboxContext } from "../../sandbox/types.js";
-import { detectAndLoadPromptImages } from "./images.js";
-import { readPersistedMediaImageLayout } from "./prompt-image-metadata.js";
-import type { EmbeddedRunAttemptParams } from "./types.js";
-
-type PromptExecutionAttempt = Pick<
-  EmbeddedRunAttemptParams,
-  | "config"
-  | "imageOrder"
-  | "images"
-  | "media"
-  | "model"
-  | "sessionFile"
-  | "sessionKey"
-  | "sessionTarget"
-  | "userTurnTranscriptRecorder"
->;
-type PromptImageResult = Awaited<ReturnType<typeof detectAndLoadPromptImages>>;
-
-function emptyPromptImages(): PromptImageResult {
-  return {
-    images: [],
-    imageFactIndexes: [],
-    detectedRefs: [],
-    failedMediaCount: 0,
-    loadedCount: 0,
-    skippedCount: 0,
-  };
-}
-
-export async function prepareEmbeddedAttemptPromptExecution(input: {
-  attempt: PromptExecutionAttempt;
-  effectiveFsWorkspaceOnly: boolean;
-  effectiveWorkspace: string;
-  prompt: string;
-  sandbox?: SandboxContext | null;
-  skipPromptSubmission: boolean;
-}): Promise<PromptImageResult> {
-  if (input.skipPromptSubmission) {
-    return emptyPromptImages();
-  }
-
-  const { attempt } = input;
-  const persistedMessage =
-    attempt.userTurnTranscriptRecorder?.message ??
-    (await attempt.userTurnTranscriptRecorder?.resolveMessage());
-  const persistedMedia = persistedMessage ? (readPersistedMediaFacts(persistedMessage) ?? []) : [];
-
-  return await detectAndLoadPromptImages({
-    prompt: input.prompt,
-    workspaceDir: input.effectiveWorkspace,
-    model: attempt.model,
-    existingImages: attempt.images,
-    imageOrder: attempt.imageOrder,
-    media: persistedMedia.length > 0 ? persistedMedia : attempt.media,
-    mediaImageLayout: persistedMessage
-      ? readPersistedMediaImageLayout(persistedMessage)
-      : undefined,
-    maxBytes: MAX_IMAGE_BYTES,
-    maxDimensionPx: resolveImageSanitizationLimits(attempt.config).maxDimensionPx,
-    workspaceOnly: input.effectiveFsWorkspaceOnly,
-    sandbox:
-      input.sandbox?.enabled && input.sandbox.fsBridge
-        ? { root: input.sandbox.workspaceDir, bridge: input.sandbox.fsBridge }
-        : undefined,
-  });
-}
