+++ src/agents/embedded-agent-runner/run/attempt-prompt-submit.ts
@@ -1,28 +1,47 @@
 /**
- * Submits one prepared prompt while owning provider transforms and cleanup.
+ * Submits or skips the prompt after build/preflight and before stream execution.
+ * It may assume prompt context is assembled and admission state is published.
  */
+import { MAX_IMAGE_BYTES } from "@openclaw/media-core/constants";
 import type { StreamFn } from "openclaw/plugin-sdk/agent-core";
 import type { ImageContent } from "../../../llm/types.js";
+import { readPersistedMediaFacts } from "../../../media/media-facts.js";
 import type { createTrajectoryRuntimeRecorder } from "../../../trajectory/runtime.js";
+import { resolveImageSanitizationLimits } from "../../image-sanitization.js";
 import type { AgentMessage } from "../../runtime/index.js";
+import type { SandboxContext } from "../../sandbox/types.js";
+import type { AgentSession } from "../../sessions/index.js";
 import { ackPendingAgentSteeringItems } from "../../subagent-registry.js";
 import { normalizeAssistantReplayContent } from "../replay-history.js";
 import { updateActiveEmbeddedRunSnapshot } from "../runs.js";
-import type {
-  getEmbeddedSessionPromptState,
-  ToolResultPromptProjectionState,
+import {
+  type getEmbeddedSessionPromptState,
+  type ToolResultPromptProjectionState,
+  hasSessionUserTurnBeenSent,
+  markSessionUserTurnsSent,
 } from "../session-prompt-state.js";
-import { hasSessionUserTurnBeenSent, markSessionUserTurnsSent } from "../session-prompt-state.js";
 import { truncateOversizedToolResultsInMessages } from "../tool-result-truncation.js";
 import { snapshotRecentMessages } from "./attempt-context-summary.js";
 import {
   installModelPromptTransform,
   installRuntimeContextMessageForPrompt,
 } from "./attempt.llm-boundary.js";
+import {
+  isSessionsYieldAbortError,
+  persistSessionsYieldContextMessage,
+  stripSessionsYieldArtifacts,
+  waitForSessionsYieldAbortSettle,
+} from "./attempt.sessions-yield.js";
+import { detectAndLoadPromptImages } from "./images.js";
 import { wrapStreamFnWithMessageTransform } from "./message-transform-stream-wrapper.js";
+import { isMidTurnPrecheckSignal, type MidTurnPrecheckRequest } from "./midturn-precheck.js";
+import { readPersistedMediaImageLayout } from "./prompt-image-metadata.js";
 import type { RuntimeContextCustomMessage } from "./runtime-context-prompt.js";
 import type { EmbeddedRunAttemptParams } from "./types.js";
 
+/**
+ * Submits one prepared prompt while owning provider transforms and cleanup.
+ */
 type PromptSubmissionSession = {
   messages: AgentMessage[];
   agent: {
@@ -166,4 +185,168 @@ export async function submitEmbeddedAttemptPrompt(input: {
     cleanupProviderPromptHistoryTransform();
     cleanupModelPromptTransform();
   }
+}
+
+type PromptSubmissionSkipReason = "blank_user_prompt" | "empty_prompt_history_images";
+
+/** Classifies prompt submissions that have no visible current-turn content. */
+export function resolvePromptSubmissionSkipReason(params: {
+  prompt: string;
+  messages: readonly unknown[];
+  imageCount: number;
+  runtimeOnly?: boolean;
+}): PromptSubmissionSkipReason | null {
+  if (params.prompt.trim().length > 0 || params.imageCount > 0) {
+    return null;
+  }
+  return params.messages.some(hasVisiblePromptHistory)
+    ? "blank_user_prompt"
+    : "empty_prompt_history_images";
+}
+
+function hasVisiblePromptHistory(message: unknown): boolean {
+  if (!message || typeof message !== "object") {
+    return false;
+  }
+  const record = message as { role?: unknown; content?: unknown };
+  if (record.role !== "user" && record.role !== "assistant") {
+    return false;
+  }
+  return hasNonEmptyContent(record.content);
+}
+
+function hasNonEmptyContent(content: unknown): boolean {
+  if (typeof content === "string") {
+    return content.trim().length > 0;
+  }
+  if (Array.isArray(content)) {
+    return content.some(hasNonEmptyContent);
+  }
+  if (!content || typeof content !== "object") {
+    return false;
+  }
+  const record = content as { text?: unknown; content?: unknown };
+  return hasNonEmptyContent(record.text) || hasNonEmptyContent(record.content);
+}
+
+/** Classifies prompt failures and performs yield or mid-turn recovery. */
+type PromptErrorAttempt = Pick<EmbeddedRunAttemptParams, "runId" | "sessionId">;
+type WithOwnedTranscriptWrite = <T>(operation: () => Promise<T> | T) => Promise<T>;
+
+type EmbeddedAttemptPromptErrorOutcome = {
+  promptFailure?: {
+    error: unknown;
+    source: "prompt";
+  };
+};
+
+export async function handleEmbeddedAttemptPromptError(input: {
+  activeSession: AgentSession;
+  attempt: PromptErrorAttempt;
+  error: unknown;
+  handleMidTurnPrecheckRequest: (request: MidTurnPrecheckRequest) => void;
+  markYieldAborted: () => void;
+  releaseLeasedSteering: (error?: unknown) => void;
+  withOwnedTranscriptWrite: WithOwnedTranscriptWrite;
+  yieldAbortSettled: Promise<void> | null;
+  yieldDetected: boolean;
+  yieldMessage: string | null;
+}): Promise<EmbeddedAttemptPromptErrorOutcome> {
+  input.releaseLeasedSteering(input.error);
+  const yieldAborted = input.yieldDetected && isSessionsYieldAbortError(input.error);
+  if (yieldAborted) {
+    // Publish terminal state before fallible recovery so outer cleanup still recognizes the yield.
+    input.markYieldAborted();
+    await waitForSessionsYieldAbortSettle({
+      settlePromise: input.yieldAbortSettled,
+      runId: input.attempt.runId,
+      sessionId: input.attempt.sessionId,
+    });
+    await input.withOwnedTranscriptWrite(async () => {
+      stripSessionsYieldArtifacts(input.activeSession);
+      if (input.yieldMessage) {
+        await persistSessionsYieldContextMessage(input.activeSession, input.yieldMessage);
+      }
+    });
+    return {};
+  }
+
+  if (isMidTurnPrecheckSignal(input.error)) {
+    const request = input.error.request;
+    await input.withOwnedTranscriptWrite(() => {
+      input.handleMidTurnPrecheckRequest(request);
+    });
+    return {};
+  }
+
+  return {
+    promptFailure: {
+      error: input.error,
+      source: "prompt",
+    },
+  };
+}
+
+/** Prepares prompt-lock ownership and prompt-local images for submission. */
+type PromptExecutionAttempt = Pick<
+  EmbeddedRunAttemptParams,
+  | "config"
+  | "imageOrder"
+  | "images"
+  | "media"
+  | "model"
+  | "sessionFile"
+  | "sessionKey"
+  | "sessionTarget"
+  | "userTurnTranscriptRecorder"
+>;
+type PromptImageResult = Awaited<ReturnType<typeof detectAndLoadPromptImages>>;
+
+function emptyPromptImages(): PromptImageResult {
+  return {
+    images: [],
+    imageFactIndexes: [],
+    detectedRefs: [],
+    failedMediaCount: 0,
+    loadedCount: 0,
+    skippedCount: 0,
+  };
+}
+
+export async function prepareEmbeddedAttemptPromptExecution(input: {
+  attempt: PromptExecutionAttempt;
+  effectiveFsWorkspaceOnly: boolean;
+  effectiveWorkspace: string;
+  prompt: string;
+  sandbox?: SandboxContext | null;
+  skipPromptSubmission: boolean;
+}): Promise<PromptImageResult> {
+  if (input.skipPromptSubmission) {
+    return emptyPromptImages();
+  }
+
+  const { attempt } = input;
+  const persistedMessage =
+    attempt.userTurnTranscriptRecorder?.message ??
+    (await attempt.userTurnTranscriptRecorder?.resolveMessage());
+  const persistedMedia = persistedMessage ? (readPersistedMediaFacts(persistedMessage) ?? []) : [];
+
+  return await detectAndLoadPromptImages({
+    prompt: input.prompt,
+    workspaceDir: input.effectiveWorkspace,
+    model: attempt.model,
+    existingImages: attempt.images,
+    imageOrder: attempt.imageOrder,
+    media: persistedMedia.length > 0 ? persistedMedia : attempt.media,
+    mediaImageLayout: persistedMessage
+      ? readPersistedMediaImageLayout(persistedMessage)
+      : undefined,
+    maxBytes: MAX_IMAGE_BYTES,
+    maxDimensionPx: resolveImageSanitizationLimits(attempt.config).maxDimensionPx,
+    workspaceOnly: input.effectiveFsWorkspaceOnly,
+    sandbox:
+      input.sandbox?.enabled && input.sandbox.fsBridge
+        ? { root: input.sandbox.workspaceDir, bridge: input.sandbox.fsBridge }
+        : undefined,
+  });
 }
